Chatgpt/requirements.txt
aiofiles==23.1.0
aiogram==2.23.1
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.19.0
aiosqlite3==0.3.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
blinker==1.7.0
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
distro==1.8.0
Django==5.0.1
Flask==3.0.0
frozenlist==1.4.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==1.0.2
Hypercorn==0.16.0
hyperframe==5.2.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
magic-filter==1.0.12
MarkupSafe==2.1.3
multidict==6.0.4
nest-asyncio==1.6.0
netaddr==0.10.1
numpy==1.26.3
openai==1.6.1
pillow==10.2.0
priority==2.0.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
pytz==2023.3.post1
Quart==0.19.4
requests==2.31.0
rfc3986==1.5.0
schedule==1.2.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
termcolor==2.4.0
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.1.0
uuid==1.30
waitress==3.0.0
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.3
yookassa==3.0.0
zope.interface==6.1
